Create Grade Analysis
This program uses numpy functions to calculate the mean, median, and standard deviation of the grades. It also finds the maximum and minimum grades, sorts the grades in ascending order, finds the index of the highest grade in the array. It also counts the number of students who scored above 90, calculates the percentage of students who scored above 90, calculates the percentage of students who scored below 75 and extracts all the grades above 90 and put them in a new array called "high_performers".